+++ app/Http/Controllers/Admin/DepartmentController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+
 use App\Exports\DepartmentsExport;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\DepartmentRequest;
@@ -8,13 +9,9 @@ use App\Imports\DepartmentsImport;
 use App\Models\Department;
 use Illuminate\Http\Request;
 use Excel;
+
 class DepartmentController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index(Request $request)
     {
         try {
@@ -32,11 +29,6 @@ class DepartmentController extends Controller
         }
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         return view('admin.department.create-edit');
@@ -45,9 +37,9 @@ class DepartmentController extends Controller
     public function store(DepartmentRequest $request)
     {
         try {
-            $data = $request->except('_token');
+            $data = $request->all();
             Department::create($data);
-            return redirect()->route('admin.departments.index')->withSuccess('Department created successfully.');
+            return redirect()->route('admin.departments.index', qArray())->withSuccess('Department created successfully.');
         } catch (\Exception $e) {
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
         }
@@ -55,20 +47,30 @@ class DepartmentController extends Controller
 
     public function show(Department $department)
     {
+        if (empty($department)) {
+            return redirect()->route('admin.departments.index');
+        }
         return view('admin.department.details', compact('department'));
     }
 
     public function edit(Department $department)
     {
-        return view('admin.department.create-edit',compact('department'));
+        if (empty($department)) {
+            return redirect()->route('admin.departments.index');
+        }
+        return view('admin.department.create-edit', compact('department'));
     }
 
     public function update(DepartmentRequest $request, Department $department)
     {
         try {
-            $data = $request->except('_token');
+            if (empty($department)) {
+                return redirect()->route('admin.departments.index');
+            }
+
+            $data = $request->all();
             $department->update($data);
-            return redirect()->back()->withSuccess('Department updated successfully.');
+            return redirect()->route('admin.departments.index', qArray())->withSuccess('Department updated successfully.');
         } catch (\Exception $e) {
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
         }
@@ -77,14 +79,18 @@ class DepartmentController extends Controller
     public function destroy(Department $department)
     {
         try {
+            if (empty($department)) {
+                return redirect()->route('admin.departments.index', qArray());
+            }
             $department->delete();
-            return redirect()->back()->withSuccess('Department trashed successfully.');
+            return redirect()->route('admin.departments.index', qArray())->withSuccess('Department trashed successfully.');
         } catch (\Exception $e) {
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
         }
     }
 
-    public function ImportExcel(Request $request) {
+    public function ImportExcel(Request $request)
+    {
         //Validation
         $this->validate($request, [
             'file' => 'required|mimes:xls,xlsx',
