+++ app/Http/Controllers/Admin/InventoryController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+
 use App\Exports\InventoryExport;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\InventoryRequest;
@@ -12,12 +13,13 @@ use App\User;
 use Illuminate\Http\Request;
 use Excel;
 use SimpleSoftwareIO\QrCode\Facades\QrCode;
+
 class InventoryController extends Controller
 {
     public function index(Request $request)
     {
         try {
-            $categoryData = Category::where('type','!=', 'Stock')->where('parent_id', 0)->with('nested')->get();
+            $categoryData = Category::where('type', '!=', 'Stock')->where('parent_id', 0)->with('nested')->get();
             $sql = Inventory::orderBy('created_at', 'ASC');
             if ($request->q) {
                 $sql->where(function ($q) use ($request) {
@@ -37,39 +39,8 @@ class InventoryController extends Controller
             if ($request->to) {
                 $sql->whereDate('purchase_date', '<=', $request->to);
             }
-            $inventories = $sql->paginate(1);
+            $inventories = $sql->paginate(50);
             return view('admin.inventory.index', compact('categoryData', 'inventories'));
-        } catch (\Exception $e) {
-            return redirect()->back()->withErrors(['error' => $e->getMessage()]);
-        }
-    }
-    public function qrCodeList()
-    {
-        try {
-            $inventories = Inventory::latest()->paginate(10);
-            return view('admin.inventory.qr-code-list', compact('inventories'));
-        } catch (\Exception $e) {
-            return redirect()->back()->withErrors(['error' => $e->getMessage()]);
-        }
-    }
-    public function summary()
-    {
-        try {
-
-            $categories_head = Category::where('type','!=', 'Stock')->with(['inventories' => function($q) {
-                $q->where('location','hq');
-            }])
-                ->get();
-
-            $categories_gs1 = Category::where('type','!=', 'Stock')->with(['inventories' => function($q) {
-                $q->where('location','gs1');
-            }])
-                ->get();
-            $categories_gs2 = Category::where('type','!=', 'Stock')->with(['inventories' => function($q) {
-                $q->where('location','gs2');
-            }])
-                ->get();
-            return view('admin.inventory.summary', compact( 'categories_head', 'categories_gs1', 'categories_gs2'));
         } catch (\Exception $e) {
             return redirect()->back()->withErrors(['error' => $e->getMessage()]);
         }
@@ -79,44 +50,44 @@ class InventoryController extends Controller
     {
         $users = User::get();
         $departments = Department::get();
-        $categoryData = Category::where('type','!=', 'Stock')->where('parent_id', 0)->with('nested')->get();
+        $categoryData = Category::where('type', '!=', 'Stock')->where('parent_id', 0)->with('nested')->get();
         return view('admin.inventory.create-edit', compact('categoryData', 'users', 'departments'));
-    }
-    public function code()
-    {
-        return view('admin.inventory.qr');
     }
 
     public function store(InventoryRequest $request)
     {
-            $data = $request->except('_token');
-            Inventory::create($data);
-            return redirect()->route('admin.inventories.index', qArray())->withSuccess('Inventory created successfully.');
+        $data = $request->all();
+        Inventory::create($data);
+        return redirect()->route('admin.inventories.index', qArray())->withSuccess('Inventory created successfully.');
 
     }
 
     public function show(Inventory $inventory)
     {
+        if (empty($inventory)) {
+            return redirect()->route('admin.inventories.index');
+        }
         return view('admin.inventory.show', compact('inventory'));
     }
-    public function showQrDetails(Inventory $inventory)
-    {
-        return view('admin.inventory.qr-code-details', compact('inventory'));
-    }
-
 
     public function edit(Inventory $inventory)
     {
+        if (empty($inventory)) {
+            return redirect()->route('admin.inventories.index');
+        }
         $users = User::get();
         $departments = Department::get();
-        $categoryData = Category::where('type','!=', 'Stock')->where('parent_id', 0)->with('nested')->get();
-        return view('admin.inventory.create-edit', compact('inventory','categoryData', 'users', 'departments'));
+        $categoryData = Category::where('type', '!=', 'Stock')->where('parent_id', 0)->with('nested')->get();
+        return view('admin.inventory.create-edit', compact('inventory', 'categoryData', 'users', 'departments'));
     }
 
     public function update(InventoryRequest $request, Inventory $inventory)
     {
         try {
-            $data = $request->except('_token');
+            if (empty($inventory)) {
+                return redirect()->route('admin.inventories.index');
+            }
+            $data = $request->all();
             $inventory->update($data);
             return redirect()->route('admin.inventories.index', qArray())->withSuccess('Inventory updated successfully.');
         } catch (\Exception $e) {
@@ -127,6 +98,9 @@ class InventoryController extends Controller
     public function destroy(Inventory $inventory)
     {
         try {
+            if (empty($inventory)) {
+                return redirect()->route('admin.inventories.index');
+            }
             $inventory->delete();
             return redirect()->route('admin.inventories.index', qArray())->withSuccess('Inventory trashed successfully.');
         } catch (\Exception $e) {
@@ -134,7 +108,51 @@ class InventoryController extends Controller
         }
     }
 
-    public function ImportExcel(Request $request) {
+    public function qrCodeList()
+    {
+        try {
+            $inventories = Inventory::latest()->paginate(10);
+            return view('admin.inventory.qr-code-list', compact('inventories'));
+        } catch (\Exception $e) {
+            return redirect()->back()->withErrors(['error' => $e->getMessage()]);
+        }
+    }
+
+    public function summary()
+    {
+        try {
+
+            $data['categories_head'] = Category::where('type', '!=', 'Stock')->with(['inventories' => function ($q) {
+                $q->where('location', 'hq');
+            }])
+                ->get();
+            $data['categories_gs1'] = Category::where('type', '!=', 'Stock')->with(['inventories' => function ($q) {
+                $q->where('location', 'gs1');
+            }])
+                ->get();
+            $data['categories_gs2'] = Category::where('type', '!=', 'Stock')->with(['inventories' => function ($q) {
+                $q->where('location', 'gs2');
+            }])
+                ->get();
+            return view('admin.inventory.summary', $data);
+        } catch (\Exception $e) {
+            return redirect()->back()->withErrors(['error' => $e->getMessage()]);
+        }
+    }
+
+
+    public function code()
+    {
+        return view('admin.inventory.qr');
+    }
+
+    public function showQrDetails(Inventory $inventory)
+    {
+        return view('admin.inventory.qr-code-details', compact('inventory'));
+    }
+
+    public function ImportExcel(Request $request)
+    {
         //Validation
         $this->validate($request, [
             'file' => 'required|mimes:xls,xlsx',
@@ -155,7 +173,8 @@ class InventoryController extends Controller
         return Excel::download(new InventoryExport, 'inventory-collection.xlsx');
     }
 
-    public function categoryInventory(Request $request, $id) {
+    public function categoryInventory($id)
+    {
         $data['category'] = $category = Category::where('id', $id)->find($id);
         $data['inventories'] = Inventory::where('category_id', $category->id)->paginate(15);
         return view('admin.inventory.summary-show', compact('data', 'category'));
